src/madsci_common/pyproject.toml
[project]
name = "madsci.common"
version = "0.6.2"
description = "The Modular Autonomous Discovery for Science (MADSci) Common Definitions and Utilities."
authors = [
    {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
    {name = "Ryan D. Lewis", email = "ryan.lewis@anl.gov"},
    {name = "Casey Stone", email = "cstone@anl.gov"},
    {name = "Doga Ozgulbas", email = "dozgulbas@anl.gov"},
]
requires-python = ">=3.10.0"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.10",
    "PyYAML>=6.0.2",
    "sqlmodel>=0.0.22",
    "python-ulid[pydantic]>=3.0.0",
    "aenum>=3.1.15",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "pymongo>=4.10.1",
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.115.4",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.17",
    "pydantic-extra-types>=2.10.2",
    "semver>=3.0.4",
    "multiprocess >= 0.70.17",
    "pydantic-settings>=2.9.1",
    "pydantic-settings-export>=1.0.2",
    "regex>=2025.7.34",
    "classy-fastapi>=0.7.0",
    "httpx>=0.28.1",
    "click>=8.1.0",
    "rich>=13.0.0",
    "opentelemetry-api>=1.20.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

otel-instrumentation = [
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-asgi>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
]

[project.scripts]
madsci-backup = "madsci.common.backup_tools.cli:main"
madsci-mongodb-backup = "madsci.common.backup_tools.cli:main_mongodb_backup"
madsci-postgres-backup = "madsci.common.backup_tools.postgres_cli:main_postgres_backup"

[project.urls]
Homepage = "https://github.com/AD-SDL/MADSci"

######################
# Build Info + Tools #
######################

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["madsci/"]
